feat(openapi): add source of OpenAPI from environment
