+++ zoo/repos/utils.py
@@ -1,9 +1,23 @@
 import importlib
+import json
 import tarfile
 import tempfile
 from pathlib import Path
 
+import redis
+import structlog
+from django.core.serializers.json import DjangoJSONEncoder
+from django.http import JsonResponse
+from pkg_resources import ResolutionError
+from prance import ResolvingParser, ValidationError
+from yaml.composer import ComposerError
+from yaml.scanner import ScannerError
+
 from .exceptions import MissingFilesError, RepositoryNotFoundError
+
+OPENAPI_SCAN_EXCLUDE = ["k8s", "test", ".gitlab", ".github"]
+
+log = structlog.get_logger()
 
 
 def get_scm_module(provider):
@@ -38,3 +52,52 @@ def download_repository(repository, fake_dir, sha=None):
             tar.extractall(fake_dir)
 
     return Path(fake_dir) / inner_folder
+
+
+def _parse_file(path, base=None):
+    try:
+        parser = ResolvingParser(str(path))
+        return parser.specification
+    except (
+        AssertionError,
+        AttributeError,
+        ComposerError,
+        FileNotFoundError,
+        ResolutionError,
+        ScannerError,
+        UnicodeDecodeError,
+        ValidationError,
+    ) as err:
+        log.info(
+            "repos.views.openapi.invalid", path=str(path.relative_to(base)), error=err
+        )
+
+
+def openapi_definition(request, repository):
+
+    redis_conn = redis.get_connection()
+    key = f"openapi-{repository.provider}-{repository.remote_id}"
+
+    if redis_conn.exists(key) and "force" not in request.GET:
+        return JsonResponse(json.loads(redis_conn.get(key)), safe=False)
+
+    specs = []
+
+    with tempfile.TemporaryDirectory() as repo_dir:
+        try:
+            repo_path = download_repository(repository, repo_dir)
+        except (MissingFilesError, RepositoryNotFoundError) as err:
+            log.info("repos.views.openapi.git_error", repo=repository, error=err)
+            return JsonResponse({"error": "error downloading repository"}, status=500)
+
+        for ext in ("json", "yml", "yaml"):
+            for path in repo_path.glob(f"**/*.{ext}"):
+                if any(directory in str(path) for directory in OPENAPI_SCAN_EXCLUDE):
+                    continue
+
+                specs.append(_parse_file(path, repo_path))
+
+    specs = list(filter(None, specs))
+
+    redis_conn.set(key, json.dumps(specs, cls=DjangoJSONEncoder), ex=1800)
+    return JsonResponse(specs, safe=False)
