+++ zoo/services/templates/services/service_form.html
@@ -75,6 +75,10 @@
       <label>{{ form.service_urls.label }}</label>
       {{ form.service_urls }}
     </div>
+    <div class="field service-urls">
+      <label>{{ form.open_api_url.label }}</label>
+      {{ form.open_api_url }}
+    </div>
   </div>
   {% endfor %}
 </div>
